use a specific controller for user self update
