+++ templates/app/controllers/chaltron/users_controller.rb
@@ -3,7 +3,7 @@ module Chaltron
     include SortAndPaginate
     helper_method :filter_provider, :filter_activity
     before_action :authenticate_user!
-    load_and_authorize_resource except: %i[self_show self_edit self_update change_password]
+    load_and_authorize_resource
 
     respond_to :html
     default_log_category :user_admin
@@ -33,27 +33,17 @@ module Chaltron
 
     def edit; end
 
-    def self_show; end
-
-    def self_edit; end
-
-    def change_password
-      redirect_to root_path unless current_user.provider.nil?
-    end
-
     def create
       if @user.save
         flash[:notice] = I18n.t('chaltron.users.created')
         info I18n.t('chaltron.logs.users.created', current: current_user.display_name, user: @user.display_name)
-        redirect_to chaltron_users_path
-      else
-        render :new, status: :unprocessable_entity
       end
+      respond_with(@user)
     end
 
     def update
       flash[:notice] = I18n.t('chaltron.users.updated') if @user.update(update_params)
-      redirect_to @user
+      respond_with(@user)
     end
 
     def enable
@@ -66,30 +56,14 @@ module Chaltron
       redirect_to(@user, notice: I18n.t('chaltron.users.disabled'))
     end
 
-    def self_update
-      change_pwd = self_update_params.include?(:password)
-      if current_user.update(self_update_params)
-        # to mantain session after password change
-        bypass_sign_in(current_user) if change_pwd
-        flash[:notice] = I18n.t('chaltron.users.self_updated')
-        redirect_to action: :self_show
-      else
-        back = :self_edit
-        back = :change_password if change_pwd
-        render back, status: :unprocessable_entity
-      end
-    end
-
     def destroy
-      options = { status: :see_other }
       if current_user == @user
-        options[:alert] = I18n.t('chaltron.users.cannot_self_destroy')
+        redirect_to({ action: :index }, alert: I18n.t('chaltron.users.cannot_self_destroy'))
       else
         info I18n.t('chaltron.logs.users.destroyed', current: current_user.display_name, user: @user.display_name)
         @user.destroy
-        options[:notice] = I18n.t('chaltron.users.deleted')
+        respond_with(@user)
       end
-      redirect_to({ action: :index }, options)
     end
 
     private
@@ -109,12 +83,6 @@ module Chaltron
 
     def update_params
       params.require(:chaltron_user).permit(role_ids: [])
-    end
-
-    def self_update_params
-      allowed = %i[fullname password password_confirmation]
-      allowed << :email if current_user.provider.nil?
-      params.require(:chaltron_user).permit(allowed)
     end
 
     def filter_params
