+++ templates/app/views/chaltron/self_user/show.html.erb
@@ -1,0 +1,60 @@
+<div class='container-fluid'>
+  <h3><%= t '.title' %></h3>
+  <div class='row'>
+
+    <div class='col-md-9'>
+      <div class='card'>
+        <div class='card-header'><%= t '.headers.details' %></div>
+        <ul class='list-group list-group-flush'>
+          <li class='list-group-item'>
+            <span><%= Chaltron::User.human_attribute_name(:username) %>: </span>
+            <strong><%= current_user.username %></strong>
+          </li>
+          <li class='list-group-item'>
+            <span><%= Chaltron::User.human_attribute_name(:fullname) %>: </span>
+            <strong><%= current_user.fullname %></strong>
+          </li>
+          <li class='list-group-item'>
+            <span><%= Chaltron::User.human_attribute_name(:email) %>: </span>
+            <strong><%= current_user.email %></strong>
+          </li>
+          <% unless current_user.department.nil? %>
+            <li class='list-group-item'>
+              <span><%= Chaltron::User.human_attribute_name(:department) %>: </span>
+              <strong><%= current_user.department %> </strong>
+            </li>
+          <% end %>
+        </ul>
+      </div>
+    </div>
+
+    <div class='col-md-3'>
+      <div class='card'>
+        <div class='card-header'><%= t '.headers.activity' %></div>
+        <ul class='list-group list-group-flush'>
+          <% unless current_user.last_sign_in_at.nil? %>
+            <li class='list-group-item'>
+              <span><%= Chaltron::User.human_attribute_name(:last_sign_in_at) %>: </span>
+              <strong><%= l current_user.last_sign_in_at, format: :short %> </strong>
+            </li>
+          <% end %>
+          <% unless current_user.last_sign_in_ip.nil? %>
+            <li class='list-group-item'>
+              <span><%= Chaltron::User.human_attribute_name(:last_sign_in_ip) %>: </span>
+              <strong><%= current_user.last_sign_in_ip %> </strong>
+            </li>
+          <% end %>
+        </ul>
+      </div>
+    </div>
+  </div>
+
+  <hr>
+
+  <%= link_to chaltron_self_user_edit_path, class: 'btn btn-primary' do %>
+    <%= icon :fas, 'user-edit', t('.edit') %>
+  <% end %>
+  <%= link_to chaltron_self_user_change_password_path, class: 'btn btn-primary' do %>
+    <%= icon :far, :edit, t('.change_password') %>
+  <% end if current_user.provider.nil? %>
+</div>
