+++ templates/config/locales/views/chaltron.it.yml
@@ -27,6 +27,7 @@ it:
         title: Elenco dei log
       logs:
         search: Cerca
+        sorted_by: Ordinamento per
         category_filter: Filtra per categoria
         severity_filter: Filtra per severità
         reset_filters: Resetta filtri
@@ -34,6 +35,8 @@ it:
       show:
         headers:
           details: Dettagli
+      log:
+        time_ago: '%{time} fa'
     users:
       index:
         title: Elenco degli utenti
@@ -57,12 +60,6 @@ it:
         submit_text: Crea utente locale
       edit:
         title: Modifica %{user}
-      self_edit:
-        title: Modifica dati personali
-        submit: Invia
-      change_password:
-        title: Cambia password
-        submit: Invia
       form:
         cancel_text: Annulla
       show:
@@ -74,13 +71,20 @@ it:
         disable: Disabilita
         destroy: Cancella
         destroy_confirm: Sei sicuro di voler cancellare %{user}?
-      self_show:
+    self_user:
+      show:
         title: Dati personali
         headers:
           activity: Attività
           details: Dettagli
         edit: Modifica dati personali
         change_password: Cambia password
+      edit:
+        title: Modifica dati personali
+        submit: Invia
+      change_password:
+        title: Cambia password
+        submit: Invia
     ldap:
       search:
         title: Ricerca utenti LDAP
